Added the source bundle.
It is a bit weird to have the parent in a subfolder, and also
problematic that for some reason the maven-source-plugin adds the .class
files to the output source jar, but at least the source bundle
generation seems to be working. (Even if the embedded jars do not have
embedded sources. I think the maven-source-plugin is not properly
configured.)
Usage: `mvn install` in the parent folder.
No hard feelings if this is not a good enough option for you, this just
an attempt to fix issue #2.